refactor(webui): promote Events to top-level nav page
Move Events out of the Settings tab onto a dedicated /events page
linked from the main navbar between Tasks and Workspaces. The Routing
Rules card moves with it so events-related controls live together.
